+++ src/mailman/rest/listconf.py
@@ -24,8 +24,8 @@ from mailman.interfaces.archiver import ArchivePolicy
 from mailman.interfaces.autorespond import ResponseAction
 from mailman.interfaces.digests import DigestFrequency
 from mailman.interfaces.mailinglist import (
-    DMARCMitigateAction, IAcceptableAliasSet, IMailingList, ReplyToMunging,
-    SubscriptionPolicy)
+    DMARCMitigateAction, IAcceptableAliasSet, IMailingList,  Personalization,
+    ReplyToMunging, SubscriptionPolicy)
 from mailman.interfaces.nntp import NewsgroupModeration
 from mailman.interfaces.template import ITemplateManager
 from mailman.model.roster import RosterVisibility
@@ -147,6 +147,7 @@ def no_newlines_validator(value):
 
 ATTRIBUTES = dict(
     acceptable_aliases=AcceptableAliases(list_of_strings_validator),
+    accept_these_nonmembers=GetterSetter(list_of_strings_validator),
     admin_immed_notify=GetterSetter(as_boolean),
     admin_notify_mchanges=GetterSetter(as_boolean),
     administrivia=GetterSetter(as_boolean),
@@ -168,14 +169,14 @@ ATTRIBUTES = dict(
     default_member_action=GetterSetter(enum_validator(Action)),
     default_nonmember_action=GetterSetter(enum_validator(Action)),
     description=GetterSetter(no_newlines_validator),
-    display_name=GetterSetter(str),
     digest_last_sent_at=GetterSetter(None),
     digest_send_periodic=GetterSetter(as_boolean),
     digest_size_threshold=GetterSetter(float),
     digest_volume_frequency=GetterSetter(enum_validator(DigestFrequency)),
     digests_enabled=GetterSetter(as_boolean),
-    dmarc_mitigate_action=GetterSetter(
-        enum_validator(DMARCMitigateAction)),
+    display_name=GetterSetter(str),
+    discard_these_nonmembers=GetterSetter(list_of_strings_validator),
+    dmarc_mitigate_action=GetterSetter(enum_validator(DMARCMitigateAction)),
     dmarc_mitigate_unconditionally=GetterSetter(as_boolean),
     dmarc_moderation_notice=GetterSetter(str),
     dmarc_wrapped_message_text=GetterSetter(str),
@@ -184,37 +185,42 @@ ATTRIBUTES = dict(
     fqdn_listname=GetterSetter(None),
     gateway_to_mail=GetterSetter(as_boolean),
     gateway_to_news=GetterSetter(as_boolean),
+    hold_these_nonmembers=GetterSetter(list_of_strings_validator),
     include_rfc2369_headers=GetterSetter(as_boolean),
     info=GetterSetter(str),
     join_address=GetterSetter(None),
     last_post_at=GetterSetter(None),
     leave_address=GetterSetter(None),
-    list_name=GetterSetter(None),
     linked_newsgroup=GetterSetter(str),
+    list_name=GetterSetter(None),
     mail_host=GetterSetter(None),
-    moderator_password=GetterSetter(password_bytes_validator),
     max_message_size=GetterSetter(integer_ge_zero_validator),
+    max_num_recipients=GetterSetter(integer_ge_zero_validator),
+    max_days_to_hold=GetterSetter(integer_ge_zero_validator),
     member_roster_visibility=GetterSetter(enum_validator(RosterVisibility)),
+    moderator_password=GetterSetter(password_bytes_validator),
     newsgroup_moderation=GetterSetter(enum_validator(NewsgroupModeration)),
     next_digest_number=GetterSetter(None),
     nntp_prefix_subject_too=GetterSetter(as_boolean),
     no_reply_address=GetterSetter(None),
     owner_address=GetterSetter(None),
+    personalize=GetterSetter(enum_validator(Personalization)),
     post_id=GetterSetter(None),
     posting_address=GetterSetter(None),
     posting_pipeline=GetterSetter(pipeline_validator),
     preferred_language=LanguageGetterSetter(language_validator),
+    reject_these_nonmembers=GetterSetter(list_of_strings_validator),
     reply_goes_to_list=GetterSetter(enum_validator(ReplyToMunging)),
     reply_to_address=GetterSetter(str),
     request_address=GetterSetter(None),
     require_explicit_destination=GetterSetter(as_boolean),
+    respond_to_post_requests=GetterSetter(as_boolean),
     send_welcome_message=GetterSetter(as_boolean),
     subject_prefix=GetterSetter(str),
     subscription_policy=GetterSetter(enum_validator(SubscriptionPolicy)),
+    unsubscription_policy=GetterSetter(enum_validator(SubscriptionPolicy)),
     usenet_watermark=GetterSetter(None),
-    volume=GetterSetter(None),
-    respond_to_post_requests=GetterSetter(as_boolean),
-    max_num_recipients=GetterSetter(integer_ge_zero_validator),
+    volume=GetterSetter(None)
     )